Streamlit/requirements.txt
streamlit
streamlit-option-menu
streamlit-folium
pandas
numpy
scikit-learn
matplotlib
plotly
seaborn
random
python-dotenv
joblib
Pillow